requirements.txt
pandas
pyyaml
ortools
matplotlib
networkx
pyvis
plotly